created better debug output for opengl, fixed type warnings

+++ Engine/src/core/Application.cpp
@@ -35,7 +35,7 @@ namespace Engine {
 		auto previous = std::chrono::steady_clock::now();
 		while (m_running) {
 			auto now = std::chrono::steady_clock::now();
-			uint32_t timeDifferenceMicroseconds = std::chrono::duration_cast<std::chrono::microseconds>(now - previous).count();
+			uint32_t timeDifferenceMicroseconds = (uint32_t) std::chrono::duration_cast<std::chrono::microseconds>(now - previous).count();
 			previous = now;
 
 			float dt = (float)timeDifferenceMicroseconds / 1000000.0f;
